+++ ouroboros-network-protocols/testlib/Ouroboros/Network/Protocol/ChainSync/Test.hs
@@ -53,7 +53,7 @@ import Test.Data.PipeliningDepth (PipeliningDepth (..))
 
 import Test.ChainGenerators ()
 import Test.ChainProducerState (ChainProducerStateForkTest (..))
-import Test.Ouroboros.Network.Testing.Utils (prop_codec_cborM,
+import Test.Ouroboros.Network.Protocol.Utils (prop_codec_cborM,
            prop_codec_valid_cbor_encoding, splits2, splits3)
 
 import Test.QuickCheck hiding (Result)
