+++ ouroboros-network-protocols/testlib/Test/Ouroboros/Network/Protocol/Utils.hs
@@ -2,8 +2,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE NamedFieldPuns   #-}
 {-# LANGUAGE RankNTypes       #-}
--- TODO rename to `Test.Ouroboros.Network.Protocols.Utils`
-module Test.Ouroboros.Network.Testing.Utils where
+module Test.Ouroboros.Network.Protocol.Utils where
 
 import Codec.CBOR.FlatTerm qualified as CBOR
 import Codec.CBOR.Read qualified as CBOR
